Added: missing translation event
A new event is triggered when a string that is missing from the
translation cache is requested for translation.

A default missing translation event listener that will insert missing
translations in the database is added ans disabled by default. It can be
enabled by adding the following lines in  app/config.php:
hs_translation:
missing_translation:
enabled: true